Implement a polyfill for WEBGL_debug
An extension, WEBGL_debug, was proposed

https://registry.khronos.org/webgl/extensions/proposals/WEBGL_debug/

The idea is similar to webgl-lint in that you can name objects
and if you're lucky, WebGL will return error messages with those
names. I say "lucky" because it's not clear which messages will
be returned as well it's up to the GL driver itself.

On proposed advantage of WEBGL_debug is ideally the labels would
be passed on the real graphics API (OpenGL, Metal, Vulkan, D3D)
and then people using native GPU debugging tools like RenderDoc,
Metal Debugger, Pix, could see the labels.

In any case, given there's a spec it seems like it wouldn't be
that hard to emulate in webgl-lint. The advantage would be
tools like spector.js might support the standard and you'd see
the labels inside spector.js. Of course spector.js could also
decide to support the extension already exposed by webgl-lint.

In any case, I started exploring what it would take to support
WEBGL_debug as a polyfill. It doesn't seem like more than a day
of work. The API for tagging objects is identical so that's just
a rename. Added to that are `pushDebugGroup` and `popDebugGroup`
which add a label and save/restore global settings.

Messages are "emitted" vs `ext.addEventListener('message')`

Enabling/disabling is via `glEnable`, `glDisable`.

There's a few other settings were you can set a log level as in
set it to error if you want only error messages emitted but not
warnings etc...

The most work is handling the case where a native `WEBGL_debug`
extension actually exists, or a least I haven't give much thought
to how to make the 2 work together since I think webgl-lint would
want to add its messages to the stream.

Otherwise, all the webgl-lint messages would need to be categorized.

In any case, I doubt that spector.js is likely to support WEBGL_debug?
and it's unclear to me if any browser will actually support it
given it's not a tiny amount of work and given WebGL is kind of nearing
a "life support only" phase with WebGPU about to take the spotlight.
So, ATM, I don't want to waste any time